refactor(routing): move the local tool ceiling to the inner boundary (BI-A8BFEFCE)
The application-boundary guard refused the first cut, correctly. Putting
`LocalPresence` and `resolveLocalToolCeiling` in `lib/tak` forced
`lib/inference/local-model-context-reconcile.ts` to import from `tak`, and the
target DAG runs the other way: inference -> routing, tak -> inference,
actions -> everything, routing -> nothing.

Three contexts need this policy — the attachment budget in `actions`, the
posture resolver in `inference`, the fallback gate in `routing` — so it belongs
at the one boundary all three may legally reach. New pure module
`lib/routing/local-tool-ceiling.ts` owns `LOCAL_TOOL_SELECTION_CLIFF`,
`LocalPresence`, and `resolveLocalToolCeiling`.
`lib/tak/context-economy-metrics.ts` re-exports all three so its existing
consumers are untouched.

This also retires an exception rather than adding one: `fallback.ts` no longer
imports from `tak` at all, so the owned `routing -> tak` debt entry is removed.
Boundary exceptions go 49 -> 48.

Module-size ratchet: `agent-coworker.ts` is compacted back to its baseline.
`fallback.test.ts` grows 1140 -> 1178 for the measured-ceiling regression test,
which is the sanctioned "unavoidable and intentional" growth. The baseline is
hand-edited for that one file rather than re-baselined wholesale — a blanket
`--update` retightened eight files this branch never touched, which would fail
other in-flight branches for growth that has nothing to do with this change.

Verification: 400 tests across 25 files still pass after the move;
`pnpm --filter web build` clean; boundary, module-size, plan-coverage and
design-grounding guards all pass locally.

Design-Grounding-Decision: no-contract-change (module relocation to satisfy the
owned application-boundary DAG; identical behaviour, re-exported for existing
consumers, no contract or authority change)
